move reducers around and add comments
